remove unused class
