--- plugin/src/test/java/org/opensearch/ml/rest/RestOpenSearchSecureTestCase.java
@@ -1,117 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.opensearch.ml.rest;
-
-import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-import org.apache.hc.core5.http.HttpHost;
-import org.junit.After;
-import org.opensearch.client.Request;
-import org.opensearch.client.Response;
-import org.opensearch.client.RestClient;
-import org.opensearch.client.RestClientBuilder;
-import org.opensearch.common.settings.Settings;
-import org.opensearch.core.xcontent.DeprecationHandler;
-import org.opensearch.core.xcontent.MediaType;
-import org.opensearch.core.xcontent.NamedXContentRegistry;
-import org.opensearch.core.xcontent.XContentParser;
-
-/**
- * Base class for running the integration tests on a secure cluster. The plugin IT test should either extend this
- * class or create another base class by extending this class to make sure that their IT can be run on secure clusters.
- */
-public abstract class RestOpenSearchSecureTestCase extends MLCommonsRestTestCase {
-
-    private static final String PROTOCOL_HTTP = "http";
-    private static final String PROTOCOL_HTTPS = "https";
-    private static final String SYS_PROPERTY_KEY_HTTPS = "https";
-    private static final String SYS_PROPERTY_KEY_CLUSTER_ENDPOINT = "tests.rest.cluster";
-    private static final String SYS_PROPERTY_KEY_USER = "user";
-    private static final String SYS_PROPERTY_KEY_PASSWORD = "password";
-    private static final String DEFAULT_SOCKET_TIMEOUT = "60s";
-    private static final String INTERNAL_INDICES_PREFIX = ".";
-    private static String protocol;
-
-    @Override
-    protected String getProtocol() {
-        if (protocol == null) {
-            protocol = readProtocolFromSystemProperty();
-        }
-        return protocol;
-    }
-
-    private String readProtocolFromSystemProperty() {
-        final boolean isHttps = Optional.ofNullable(System.getProperty(SYS_PROPERTY_KEY_HTTPS)).map("true"::equalsIgnoreCase).orElse(false);
-        if (!isHttps) {
-            return PROTOCOL_HTTP;
-        }
-
-        // currently only external cluster is supported for security enabled testing
-        if (Optional.ofNullable(System.getProperty(SYS_PROPERTY_KEY_CLUSTER_ENDPOINT)).isEmpty()) {
-            throw new RuntimeException("cluster url should be provided for security enabled testing");
-        }
-        return PROTOCOL_HTTPS;
-    }
-
-    @Override
-    protected RestClient buildClient(Settings settings, HttpHost[] hosts) throws IOException {
-        final RestClientBuilder builder = RestClient.builder(hosts);
-        if (PROTOCOL_HTTPS.equals(getProtocol())) {
-            configureHttpsClient(builder, settings);
-        } else {
-            configureClient(builder, settings);
-        }
-
-        return builder.build();
-    }
-
-    /**
-     * wipeAllIndices won't work since it cannot delete security index. Use deleteExternalIndices instead.
-     */
-    @Override
-    protected boolean preserveIndicesUponCompletion() {
-        return true;
-    }
-
-    @After
-    public void deleteExternalIndices() throws IOException {
-        final Response response = client().performRequest(new Request("GET", "/_cat/indices?format=json" + "&expand_wildcards=all"));
-        final MediaType xContentType = MediaType.fromMediaType(response.getEntity().getContentType());
-        try (
-            final XContentParser parser = xContentType
-                .xContent()
-                .createParser(
-                    NamedXContentRegistry.EMPTY,
-                    DeprecationHandler.THROW_UNSUPPORTED_OPERATION,
-                    response.getEntity().getContent()
-                )
-        ) {
-            final XContentParser.Token token = parser.nextToken();
-            final List<Map<String, Object>> parserList;
-            if (token == XContentParser.Token.START_ARRAY) {
-                parserList = parser.listOrderedMap().stream().map(obj -> (Map<String, Object>) obj).collect(Collectors.toList());
-            } else {
-                parserList = Collections.singletonList(parser.mapOrdered());
-            }
-
-            final List<String> externalIndices = parserList
-                .stream()
-                .map(index -> (String) index.get("index"))
-                .filter(indexName -> indexName != null)
-                .filter(indexName -> !indexName.startsWith(INTERNAL_INDICES_PREFIX))
-                .collect(Collectors.toList());
-
-            for (final String indexName : externalIndices) {
-                adminClient().performRequest(new Request("DELETE", "/" + indexName));
-            }
-        }
-    }
-}
